pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "glaze-gallery"
version = "0.1.0"
description = "Scripts for syncing data for the live glaze gallery websites."
authors = ["Alex Hadley <contact@alexhadley.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytz = "^2024.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
pillow = "^11.0.0"
google-api-python-client = "^2.151.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
tqdm = "^4.66.6"

[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
types-tqdm = "^4.66.0.20240417"
pandas-stubs = "^2.2.3.241009"

[tool.mypy]
strict = true

[tool.poetry.scripts]
glaze_gallery_download = 'glaze_gallery._download:download'
